Fixed bug in scanner
If this was the last import statement:

import "package:distributed_dart/distributed_dart.dart" as dist;

We got a keyword (dist) at last and never stopped because we never
stopped before we got a new string (which could be inside the code).
Now we set nextTokenIsImportant to false after each semicolon.