Monster update!
Removed DartCode and DartCodeChild.
Added DartProgram, DependencyNode and FileNode.
Lots of refactoring of code.
Still missing to fix the documentation (lots of it is wrong or missing
now)

The most important fix is we now find files there are included multiple
times and remove them. This fix a major bug where the link component try
create links there already exists.