Full path on links in Linux = bad
Again the problem when you ask for the full path on a link you got the
location of the file the link points to.

Now we use the current working dir for the running dart process
(Directory.current.path) to build the full path to the dart process we
want to spawn.